Latest RM0314-1528-0.420.0-86a8013 on PATREON - UPD SUBGHz

+++ lib/subghz/protocols/revers_rb2.h
@@ -78,7 +78,7 @@ void subghz_protocol_decoder_revers_rb2_feed(void* context, bool level, uint32_t
  * @param context Pointer to a SubGhzProtocolDecoderRevers_RB2 instance
  * @return hash Hash sum
  */
-uint8_t subghz_protocol_decoder_revers_rb2_get_hash_data(void* context);
+uint32_t subghz_protocol_decoder_revers_rb2_get_hash_data(void* context);
 
 /**
  * Serialize data SubGhzProtocolDecoderRevers_RB2.
